// week-4/enterprise-composer-app/src/app/composer.interface.ts
/*
================================================================
    Title: composer.interface.ts
    Author: Carl Logan
    Date: 01/27/2023
    Description: Assignment 4.4 - Async Pipe.
================================================================
*/

export interface IComposer {
  composerId: number;
  fullName: string;
  genre: string;
}